+++ src/main/resources/schema.sql
@@ -3,4 +3,4 @@ CREATE TABLE IF NOT EXISTS snippets (
                                      snippet VARCHAR(1000) NOT NULL,
     snippet_date TIMESTAMP,
     view_limit INTEGER,
-    time_limit INTEGER);
+    time_limit BIGINT);
